refactor: data 패키지 common 에서 분리
