--- app/src/main/java/woowacourse/shopping/common/data/database/table/SqlTable.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.common.data.database.table
-
-import woowacourse.shopping.common.data.database.SqlColumn
-
-interface SqlTable {
-    val name: String
-    val scheme: List<SqlColumn>
-    val constraint: String
-}
