+++ app/src/main/java/woowacourse/shopping/data/database/ShoppingDBOpenHelper.kt
@@ -1,12 +1,12 @@
-package woowacourse.shopping.common.data.database
+package woowacourse.shopping.data.database
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
-import woowacourse.shopping.common.data.database.table.SqlCart
-import woowacourse.shopping.common.data.database.table.SqlProduct
-import woowacourse.shopping.common.data.database.table.SqlRecentProduct
-import woowacourse.shopping.common.data.database.table.SqlTable
+import woowacourse.shopping.data.database.table.SqlCart
+import woowacourse.shopping.data.database.table.SqlProduct
+import woowacourse.shopping.data.database.table.SqlRecentProduct
+import woowacourse.shopping.data.database.table.SqlTable
 
 class ShoppingDBOpenHelper(context: Context) :
     SQLiteOpenHelper(context, DB_NAME, null, DB_VERSION) {
