+++ app/src/main/java/woowacourse/shopping/data/database/table/SqlCart.kt
@@ -1,13 +1,13 @@
-package woowacourse.shopping.common.data.database.table
+package woowacourse.shopping.data.database.table
 
-import woowacourse.shopping.common.data.database.SqlColumn
-import woowacourse.shopping.common.data.database.SqlType
+import woowacourse.shopping.data.database.SqlColumn
+import woowacourse.shopping.data.database.SqlType
 
-object SqlRecentProduct : SqlTable {
+object SqlCart : SqlTable {
     const val TIME = "time"
     const val PRODUCT_ID = "product_id"
 
-    override val name: String = "RecentProduct"
+    override val name: String = "Cart"
     override val scheme: List<SqlColumn> = listOf(
         SqlColumn(TIME, SqlType.TEXT, "PRIMARY KEY"),
         SqlColumn(PRODUCT_ID, SqlType.INTEGER)
